spring3

+++ helloSpring/src/main/java/com/hello/project1/Emp.java
@@ -4,8 +4,9 @@ public class Emp {
 	
 	private int id;
 	private String name;
-	private float salary;
+	private int salary;
 	private String designation;
+	
 	public int getId() {
 		return id;
 	}
@@ -18,10 +19,10 @@ public class Emp {
 	public void setName(String name) {
 		this.name = name;
 	}
-	public float getSalary() {
+	public int getSalary() {
 		return salary;
 	}
-	public void setSalary(float salary) {
+	public void setSalary(int salary) {
 		this.salary = salary;
 	}
 	public String getDesignation() {
